+++ src/session_base.hpp
@@ -36,7 +36,8 @@
 #include "io_object.hpp"
 #include "pipe.hpp"
 #include "socket_base.hpp"
-#include "stream_engine.hpp"
+#include "i_engine.hpp"
+#include "msg.hpp"
 
 namespace zmq
 {
@@ -61,7 +62,7 @@ class session_base_t : public own_t, public io_object_t, public i_pipe_events
     virtual void reset ();
     void flush ();
     void rollback ();
-    void engine_error (zmq::stream_engine_t::error_reason_t reason_);
+    void engine_error (zmq::i_engine::error_reason_t reason_);
 
     //  i_pipe_events interface implementation.
     void read_activated (zmq::pipe_t *pipe_);
